test(auth): move OAuth state cookie tests
